fix(frontend): show blank selections for area and season when there is no selection
Previously, the <select> elements showed the first value in the options even though it was not selected. Now, when no option is selected, an empty option is inserted in front of the other options so that the dropdown shows that as the selection. The empty option goes away once a value is actually selected.

In the future, we will have the app set a defauly value in the URL (defaults TBD).