feat(pipeline/replica): add a credentialless mode for the replica ETL (#36)
* feat(pipeline/replica): add a credentialless mode for the replica ETL

When BigQuery credentials are unavailable, the replica ETL will attempt to infer the datasets table from the parquet files located in ./data/replica/full_area. If there are no files, credentials will still be required because there will be no data to locally process.

This is useful when running the ETL without replica BigQuery credentials. If you have access to downloaded data for the full_area extent, you can manually place the data in ./data/replica/full_area to be able to run the pipeline.

* docs(pipeline/replica): finish docstring for `read_schema`

* fix(pipleine/replica): raise an error instead of nothing in `read_schema`