feat(frontend): optimize the pipeline output data copy process and update it for the latest file names (#37)
* fix(frontend): only copy json, geojson, and deflate files from pipeline data output

This massively speeds up the copying process.

* feat(frontend): delete public data folder contents before copying new data from pipeline data output

* fix(frontend): remove empty directories before building the area and season indices

If we do not move the empty directories, the area index will include the names of the empty directories as valid areas.

* feat(frontend): support new replica etl output file names

The new names do not include the dataset name in the file name because it is already included in the parent folder name.