feat(pipeline/replica): do not re-download from replica unless data is missing
This commit adds logic to check if a datatset is already downloaded for its associated season. If it is, it will be ommited from the dataframe of tables to download.

This change makes it possible to get the newest data once it is released without needing to specify a year and quarter filter to avoid re-downloading all past data.