feat(pipeline/essential_services)!: generate separate stats files for each season-area
This ensures that running the data pipeline with a narrower set of areas or seasons does not result in loss the previous stats. An example use case is re-running the pipeline when a new season is released or adding a new area after the pipeline has been run for the other areas.