feat: add subtab to tab5 for listing all scenario features in one view (#207)
This PR adds a new subtab to tab 5 (Roads or Transit) for listing all scenario features in a single scrollable page. As part of this change, the highlight and zoom logic was extracted into a separate function that is used by the 2 subtabs.

On mobile view, the subtabs appear the exact same way as the sections subtabs on the other 4 tabs.