fix: reduce unnecessary rerenders whne search params change
Areas, seasons, and travel methods are used by the app data context. If those change whenever the URL changes - even if their respective values did not change - any part of the app reliant on `useAppData` will re-render. This change ensures that re-renders are only triggered due to areas, season, or travel methods when one of the values in the URL actually changes.