feat(frontend): use custom button and calcite panel instead of un-styleable arcgis expand widget
The recent uipdate to ArcGIS SDK for JavaScript broke our ability to override the styles of the individual components. By switching to our button component and the calcite sheet component, we can restore the old layer manager button while providing an improved experience for the layers manager. The layers manager sheet is designed to look similar to our existing options sidebar.

Fixes #263.